<fix> ID 값 Integer로 맞춤

+++ src/main/java/selecting/platform/repository/ChatRoomRepository.java
@@ -3,5 +3,5 @@ package selecting.platform.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import selecting.platform.model.ChatRoom;
 
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Integer> {
 }
